+++ laktory/models/datasources/pipelinenodedatasource.py
@@ -77,7 +77,14 @@ class PipelineNodeDataSource(BaseDataSource):
 
             if is_debug():
                 df = None
-                if self.node.output_df:
+
+                stream_to_batch = (
+                    not self.as_stream
+                    and self.node.output_df
+                    and self.node.output_df.isStreaming
+                )
+
+                if self.node.output_df and not stream_to_batch:
                     logger.info(
                         f"Reading pipeline node {self._id} from output DataFrame (DLT debug)"
                     )